backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
wsproto==1.2.0
